2nd Changes to Code
Crouch added to player movement. the Camera just moves down and enemies pretend they cant see you if they havent spotted you already

We havent decided on a theme yet so I just called the things we collect souls cause why not. Collect them all to win the game

Two enemy types added
1.  the spinning head enemy spins its head in player direction as soon as they are nearby. And chases the player if the player isnt crouching. This enemy patrols between collectable Souls
2. the second enemy finds a soul then walks in circles around it.
if the enemies touch you you lose.

theres a pause panel that looks basic asf but it works and so does the rest of the UI